requirements.txt
openai
langchain
langgraph
chromadb
tavily-python
python-dotenv
streamlit
requests
beautifulsoup4
pandas
python-docx
reportlab
PyMuPDF
google-api-python-client
google-auth
google-auth-oauthlib
spacy
python-telegram-bot==20.6
openai
httpx
SpeechRecognition
pydub
google-api-python-client
python-dotenv
